Small optimization
